+++ Content/Python/UEFN_Toolbelt/tools/__init__.py
@@ -74,3 +74,4 @@ from . import movie_render_tools
 from . import viewport_tools
 from . import actor_visibility
 from . import verse_templates
+from . import cooker_optimizer